remove redundant containers
