+++ src/routes/Admin/Players/components/DaySelector.js
@@ -1,5 +1,15 @@
+import { connect } from 'react-redux'
 import React from 'react'
 import PropTypes from 'prop-types'
+import { changeSortDay } from '../modules/players'
+
+const mapStateToProps = state => ({
+  loading: state.playersApp.loading
+})
+
+const mapDispatchToProps = {
+  changeSortDay: changeSortDay
+}
 
 const DaySelector = ({ changeSortDay }) => (
   <select className='select-day'
@@ -13,4 +23,4 @@ DaySelector.propTypes = {
   changeSortDay: PropTypes.func.isRequired,
 }
 
-export default DaySelector
+export default connect(mapStateToProps, mapDispatchToProps)(DaySelector)
