+++ src/routes/Admin/Players/components/PlayerShow.js
@@ -1,5 +1,18 @@
+import { connect } from 'react-redux'
 import React from 'react'
 import PropTypes from 'prop-types'
+import { editPlayer } from '../modules/players'
+import { calcTotals } from '../../../../services/score.service'
+
+const mapDispatchToProps = dispatch => ({
+  editPlayer: id => {
+    dispatch(editPlayer(id))
+  },
+})
+
+const mapStateToProps = (state, props) => {
+  return calcTotals(props.scores_day1, props.scores_day2)
+}
 
 const PlayerShow = ({
                       // props
@@ -68,4 +81,4 @@ PlayerShow.propTypes = {
   totalScore: PropTypes.number.isRequired,
 }
 
-export default PlayerShow
+export default connect(mapStateToProps, mapDispatchToProps)(PlayerShow)
